@@ -20,8 +20,8 @@
         }
     ],
     "require": {
-        "symfony/framework-bundle": "~2.3",
-        "symfony/form": "~2.7"
+        "symfony/framework-bundle": "~2.8|~3.0",
+        "symfony/form": "~2.8|~3.0"
     },
     "autoload": {
         "psr-4": {
